requirements.txt
bayoo-docx
lxml